Log: Refactored app structure, minor changes to avoid warnings
